Upgrade to Join-table
Changed the storage of contract parties from contractEntity to joined table of contractId and partyId.

Changed/implemented the function for Insured being added/removed from contracts.

Changed/implemented the function to remove/add contracts and create the contract with the role automaticaly.

Started implementation of stoping to store users as Policy_Owner or Insured to having a role assigned on the given contract - transition still needs to be finished.

Fully working CRD of Insurances, Contracts, Parties, just the U part left to do before final clean up of this stage.

Changed the display of Parties to show all contracts it is part of - still needs update of the details for this fact

Changed the display of Contracts to show all parties that are part of it. Also added the name of the insurance as a reference link for its details.

Changed entities to account for the updated jointable and also removed two redundant vars from entities and its DTOs.

And many other but related to these changes.