Big Refactoring 1
Shortlived cleanup of var names and comments.

Removed faulty entity vars as contractNumber, insuranceType since those are being tracked seperately now.
Updated the DTOs accordingly.

Accounted searching of parties for this, from searching based in party.contractNumber to the use of a query that get the PartyContract entity and finds the entry based on contractId and partyId.
Added the search option for parties without any contracts or insurances.
Removed redundand create and delete option for insured as this is all now being created and deleted into a contracless party.
Changed the menu accordingly for this.

Fully changed filtering of parties to account for this since partyEntity no longer holds the partyStatus/contractRole with redundancy checks.

Improved the display of party contracts and contract parties display to show their roles in the contract and insuranceType.

Updated HTML templates for all of these changes.

Fully working CRD of previous functions checked.